Addition cal and carbs

+++ SwiftProject/SwiftProject/Food.swift
@@ -4,16 +4,16 @@ struct Food: Codable, Identifiable {
     let id = UUID()
     var name: String
     var calories: Float
-    var serving_size_g: Double
-    var fat_total_g: Double
-    var fat_saturated_g: Double
-    var protein_g: Double
-    var sodium_mg: Double
-    var potassium_mg: Double
-    var cholesterol_mg: Double
-    var carbohydrates_total_g: Double
-    var fiber_g: Double
-    var sugar_g: Double
+    var serving_size_g: Float
+    var fat_total_g: Float
+    var fat_saturated_g: Float
+    var protein_g: Float
+    var sodium_mg: Float
+    var potassium_mg: Float
+    var cholesterol_mg: Float
+    var carbohydrates_total_g: Float
+    var fiber_g: Float
+    var sugar_g: Float
 }
 
 class Api : ObservableObject{
